% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{add_r_profile}
\alias{add_r_profile}
\title{Custom .Rprofile file}
\usage{
add_r_profile(..., template = ".Rprofile")
}
\arguments{
\item{...}{dots. Arguments passed to \code{usethis::use_template()}.}

\item{template}{chracter. Default value is provided.}
}
\description{
Add a \code{.Rprofile} file to the project.
}
\details{
Calls \code{usethis::use_template()}.
}
\seealso{
\url{https://github.com/r-lib/usethis/blob/master/R/template.R}.
}
