% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{add_template}
\alias{add_template}
\alias{add_r_profile}
\alias{add_task_schedule}
\title{Custom template file}
\usage{
add_template(template, ..., package = "teproj")

add_r_profile(template = ".Rprofile", ...)

add_task_schedule(..., template = "schedule-task.R")
}
\arguments{
\item{template}{chracter. Default value is provided.}

\item{...}{dots. Arguments passed to \code{usethis::use_template()}.}
}
\description{
Add a {template} file to the project.
}
\details{
Calls \code{usethis::use_template()}.
}
\seealso{
\url{https://github.com/r-lib/usethis/blob/master/R/template.R}.
}
