% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{arrange_distinctly}
\alias{arrange_distinctly}
\title{\code{dplyr::arrange()} + \code{dplyr::distinct()}}
\usage{
arrange_distinctly(data = NULL, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{dots. Bare names of columns in \code{data} on which to perform operations.}
}
\value{
data.frame
}
\description{
Shorthand for \code{dplyr} functions called consecutively.
}
\details{
None.
}
