% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ext.r
\name{create_dir}
\alias{create_dir}
\title{Create a directory.}
\usage{
create_dir(dir = paste0(getwd(), "/"), overwrite = FALSE, backup = TRUE,
  create = TRUE, ...)
}
\arguments{
\item{dir}{Character. Folder name (with a trailing slash), Default: paste0(getwd(), "/").}

\item{overwrite}{Boolean. Default: FALSE.}

\item{backup}{Boolean. Only relevant if `dir` \code{dir} exists and `overwrite == TRUE` \code{overwrite == TRUE}. Default: TRUE.}

\item{create}{Boolean. Intended to be used with a global `keep_*` parameter. Default: TRUE.}

\item{...}{Dots. Additional parameters to pass to `dir.create()`.}
}
\value{
Character. Filepath.
}
\description{
Creates the directory if it does not exist.
}
\details{
Used by other functions in this package.
}
