% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create an example project/file.}
\usage{
create_example(style = "personal", template = "analysis",
  dir = "teproj_test", ..., overwrite = FALSE)
}
\arguments{
\item{style}{Character. Valid arguments include "ercot".}

\item{template}{Character. Valid arguments include "analysis".}

\item{dir}{Character. Name of directory to create. Default is provided.}

\item{...}{Dots. Not currently used.}

\item{overwrite}{Boolean. Indicates whether to overwrite existing directory or not.}
}
\value{
Character. Directory filepath.
}
\description{
Creates an example project or file.
}
\details{
TODO.
}
