% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kable.R
\name{create_kable}
\alias{create_kable}
\alias{create_kable_html}
\alias{create_kable_md}
\title{Create a directory}
\usage{
create_kable(data = NULL, n_show = 20L, show_footnote = ifelse(nrow(data)
  > n_show, TRUE, FALSE), format = "html", ..., full_width = FALSE,
  position = "center")

create_kable_html(..., format = "html")

create_kable_md(..., format = "markdown")
}
\arguments{
\item{data}{data.frame.}

\item{n_show}{integer. Number of rows to show in output.}

\item{show_footnote}{logical. Whether to show total number of rows if \code{data} is truncated.}

\item{format}{character. Passed directly to same \code{knitr::kable()} argument.}

\item{...}{dots. Not currently used.}

\item{full_width}{logical. Passed directly to same \code{kableExtra::kable_styling()} argument.}

\item{position}{character. Passed directly to same \code{kableExtra::kable_styling()} argument.}
}
\value{
kable object.
}
\description{
Creates the directory if it does not exist.
}
\details{
Used by other functions in this package.
Note that this function is probably more complex than it really should be.
Re-factoring it should be considered.
}
