% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{do_call_with}
\alias{do_call_with}
\title{\code{do.call()} with package namespace}
\source{
\url{https://stackoverflow.com/questions/10022436/do-call-in-combination-with}.
}
\usage{
do_call_with(what, args, ...)
}
\description{
Allows a package to explicitly included in \code{do.call()}
}
\details{
None.
}
