% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get_path_lazily}
\alias{.get_path_lazily}
\title{Construct a file path}
\usage{
.get_path_lazily(dir = NULL, ..., ext = NULL)
}
\arguments{
\item{dir}{character. Folder name (with a trailing slash).}

\item{...}{dots. Arguments parssed to construct name of file path without the directory nor
the extension. Uses \code{paste()} to collapse all arguments between \code{dir} and \code{ext}.}

\item{ext}{character. Bare extension (i.e. without a dot). Must be one of valid formats.}
}
\value{
character. Path.
}
\description{
Constructs a file path given a directory as the first input and an extension as the last input.
}
\details{
Assumes that the last arguments is the path extension. If \code{ext} is not
specified explicitly, it is assumed to be the last argument passed to the function.
This format allows for 'lazy' construction of the basename of the file (sans extension)
given all arguments that are not the first nor the last.
}
\keyword{internal}
