% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.sort_named_list}
\alias{.sort_named_list}
\title{Sort a named list}
\usage{
.sort_named_list(x)
}
\arguments{
\item{x}{list.}
}
\value{
list.
}
\description{
Sort a named list.
}
\details{
None.
}
\keyword{internal}
