% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.unrowname}
\alias{.unrowname}
\title{Remove rownames from data.frame}
\source{
\url{https://github.com/tidyverse/broom/blob/master/R/utilities.R}.
}
\usage{
.unrowname(data = NULL)
}
\arguments{
\item{data}{data frame}
}
\value{
data.frame
}
\description{
Remove rownames from data.frame
}
\keyword{internal}
