% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_path_safely}
\alias{get_path_safely}
\title{Construct a file path}
\usage{
get_path_safely(dir = NULL, file = NULL, ext = NULL, path = NULL)
}
\value{
character. Path.
}
\description{
Construct a file path given a directory, file, and extension.
}
\details{
This function is notable because it has a \code{path} argument
that can be used as a quick "out".
This helps with argument checking with package functions.
}
