% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_opts.R
\name{get_pkg_opts_renamed}
\alias{get_pkg_opts_renamed}
\alias{get_pkg_opts_verbose}
\title{View package options}
\usage{
get_pkg_opts_renamed(type = c("print", "render", "ggsave"),
  pkg_opts_prefix = "teproj")

get_pkg_opts_verbose(type = c("print", "render", "ggsave"),
  pkg_opts_prefix = "teproj")
}
\arguments{
\item{type}{character. Specification of which package options to extract (i.e. "print", "render", or "ggsave").}

\item{pkg_opts_prefix}{character. "teproj".}
}
\value{
list (of named characters). Renamed package options.
}
\description{
View options for relevant \code{teproj} functions.
}
\details{
Particularly useful for the \code{render_pojr_io()} function.
}
