% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{import_ext}
\alias{import_ext}
\alias{import_ext_csv}
\alias{import_ext_xlsx}
\alias{import_excel}
\alias{import_ext_rda}
\alias{import_ext_rdata}
\alias{import_ext_RData}
\alias{import_ext_rds}
\alias{import_ext_feather}
\title{Import an object.}
\usage{
import_ext(x = NULL, filename = deparse(substitute(x)),
  dir = paste0(getwd(), "/"), ext = NULL, filepath = paste0(dir, filename,
  ".", ext), ...)

import_ext_csv(...)

import_ext_xlsx(...)

import_excel(...)

import_ext_rda(...)

import_ext_rdata(...)

import_ext_RData(...)

import_ext_rds(...)

import_ext_feather(...)
}
\arguments{
\item{x}{Data.frame (or matrix) for most formats. Default: NULL.}

\item{filename}{Character. Bare filename (i.e. without folderor extension), Default: deparse(substitute(x)).}

\item{dir}{Character. Folder name (with a trailing slash), Default: paste0(getwd(), "/").}

\item{ext}{Character. Bare extension (i.e. without a dot). Must be one of valid formats. Default: NULL.}

\item{filepath}{Character. Concatenation of `filename`, `dir`, and `ext`, Default: paste0(dir, filename, ".", ext).}

\item{...}{Dots. May (or may not) be passed on to specific exporting functions.}
}
\description{
Reads in an object from a filepath.
}
\details{
None.
}
