% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_path}
\alias{import_path}
\title{Import an object}
\usage{
import_path(path = NULL, import = TRUE, return = TRUE, ...)
}
\arguments{
\item{path}{character. Concatenation of \code{file}, \code{dir}, and \code{ext},}

\item{import}{logical. Indicates whether to actually execute function.}

\item{return}{logical. Relevant ONLY if \code{export == FALSE}.
Set to \code{TRUE} in order to preview what would be rendered.}

\item{...}{dots. Arguments to pass dircetly to internally used import function.}
}
\value{
object.
}
\description{
Reads in data given a full path.
}
\details{
Works similarly to \code{import_ext} internally, but may be considered simpler.
}
