% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_path_cleanly}
\alias{import_path_cleanly}
\title{Import an object}
\usage{
import_path_cleanly(path = NULL, ...)
}
\arguments{
\item{path}{character. Concatenation of \code{file}, \code{dir}, and \code{ext},}

\item{...}{dots. Arguments to pass dircetly to internally used import function.}
}
\value{
object.
}
\description{
Reads in data given a full path and 'cleans' the names.
}
\details{
Calls \code{import_path} internally.
}
