% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_proj_io.R
\name{parse_proj_io}
\alias{parse_proj_io}
\title{Parse project input/output.}
\usage{
parse_proj_io(filepaths, dir, ..., rgx_file_io = ".R",
  rgx_input = ".*(import_ext|teutils::import_ext)",
  rgx_output = ".*(export_ext|teutils::export_ext)")
}
\arguments{
\item{filepaths}{character (vector),}

\item{dir}{character. Should not be a vector, Only used if filepaths is missing.
#' @param ... dots. Parameters passed to `list.files()`. (Technically, `pattern` could/should not be specified explicitly.}

\item{...}{dots. Additional parameters to pass to `dir.create()`.}

\item{rgx_file_io}{character. Alias to `pattern` parameter for `list.files()`. Used ONLY if `filepaths` is missing and `dir` is not.}

\item{rgx_input}{character. Regular expression to match for input functions,}

\item{rgx_output}{Chracter. Regular expression to match for out functions.,}
}
\value{
data.frame
}
\description{
Parses project to identify inputs/out variables, filenames, etc.
}
\details{
Intended to be used with `export_ext()*` and `import_ext*()` functions, but can work
if `rgx_input` and `rgx_output` are modified appropriately.
Does not currently work as intended if variables, filenames, etc. are not wrapped by `rgx_input` and `rgx_output`.
(i.e. Does not work as intended if variables, filenames, etc. is on a separate line,
such as with a pipe.)
}
