% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{pull_distinctly}
\alias{pull_distinctly}
\title{\code{dplyr::distinct()} + \code{dplyr::arrange()} + \code{dplyr::pull()}}
\usage{
pull_distinctly(data = NULL, col)
}
\arguments{
\item{data}{data.frame}

\item{col}{charactor for SE version; symbol for NSE version. Name of column in \code{data} on which to perform operations.}
}
\value{
vector.
}
\description{
Shorthand for \code{dplyr} functions called consecutively.
}
\details{
None.
}
