% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pkg_opts.R
\name{set_pkg_ggssave_opts}
\alias{set_pkg_ggssave_opts}
\title{Control package option
.}
\usage{
set_pkg_ggssave_opts(units = getOption("teproj.ggsave.units"),
  width = getOption("teproj.ggsave.width"),
  height = getOption("teproj.ggsave.height"))
}
\arguments{
\item{units}{character. \code{ggsave()} parameters.}

\item{width, height}{numerics. \code{ggsave()} parameters.}
}
\description{
Sets options for saving a \code{ggplot2} plot with \code{teproj::export()}.
}
\details{
Intended to be used as a wrapper to \code{options(...)}.
}
