% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{set_pkg_print_opts}
\alias{set_pkg_print_opts}
\title{Control package messages.}
\usage{
set_pkg_print_opts(msg = getOption("teproj.print_msg"),
  wrn = getOption("teproj.print_wrn"), err = getOption("teproj.print_err"))
}
\arguments{
\item{msg}{boolean.}

\item{wrn}{boolean.}

\item{err}{boolean.}
}
\description{
Sets options for relevant `teproj` functions.
}
\details{
Intended to be used as a wrapper to `options(...`).
}
