% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{set_pkg_render_opts}
\alias{set_pkg_render_opts}
\title{Control package render settings.}
\usage{
set_pkg_render_opts()
}
\description{
Sets options for relevant `teproj` functions.
}
\details{
Intended to be used as a wrapper to `options(...`).
}
