% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{summarise_stats_at}
\alias{summarise_stats_at}
\alias{summarise_stats}
\alias{summarise_stats_by_at}
\title{Summary statistics for a data.frame}
\usage{
summarise_stats_at(data = NULL, col = NULL, ..., na.rm = TRUE,
  tidy = FALSE)

summarise_stats(data = NULL, col, ..., tidy = FALSE)

summarise_stats_by_at(data = NULL, col = NULL, cols_grp = NULL, ...)
}
\arguments{
\item{data}{data.frame}

\item{col}{charactor for SE version; symbol for NSE version. Name of column in \code{data} on which to perform operations.}

\item{...}{dots. Arguments passed to stats functions used internally.}

\item{na.rm}{logical. Argument passed to stats function used internally.}

\item{tidy}{logical. Whether to put output in long (i.e. tidy) format.}

\item{cols_grp}{charactor for SE version; symbol for NSE version.}
}
\value{
data.frame.
}
\description{
A customized summary of a data.frame.
}
\details{
The purpose of this function is equivalent to that of similar functions in
other packages, such as \code{skimr::skim()}.
This function outputs the following \code{n}, \code{mean}, \code{median}, \code{sd},
\code{min}, \code{max}, \code{zn1}, \code{zp1}, \code{q25}, \code{q75}, \code{q05}, \code{q95}.

The \code{_at} versions of this function are SE evaluation. (i.e. They take characters
as column names.) The \code{_by} version(s) of this function allows
for groups to be specificed as an input, although this function will detect groups
and respect their integrity  (meaning that the \code{_by} version(s) are simply
provided as an alternative means).
}
\seealso{
\url{https://github.com/ropenscilabs/skimr/blob/master/R/skim.R}.
}
