% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{.use_template}
\alias{.use_template}
\alias{use_r_profile}
\alias{use_task_schedule}
\title{Custom template file}
\usage{
.use_template(template, save_as = template, package = "teproj", ...)

use_r_profile(template = ".Rprofile", save_as = template, ...)

use_task_schedule(template = "schedule-task.R", save_as = file.path("R",
  template), ...)
}
\arguments{
\item{template}{character. Name of template file. Default value is provided.}

\item{save_as}{character. Location/name of file to save. Default is provided.}

\item{package}{character. Name of package. Default is name of this package.}

\item{...}{dots. Arguments passed to \code{usethis::use_template()}.}
}
\description{
Add a {template} file to the project.
}
\details{
Calls \code{usethis::use_template()}. Really only would need to use
\code{system.file()}, \code{file.copy()}, etc. if there are no variables to be filled in the template file. (See the \code{data} argument of \code{usethis::use_template()}.
}
\seealso{
\url{https://github.com/r-lib/usethis/blob/master/R/template.R}.
}
