% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warningf}
\alias{warningf}
\title{\code{warning} + \code{sprintf}}
\source{
\url{https://stackoverflow.com/questions/9596918/r-warning-wrapper-raise-to-parent-function}.
}
\usage{
warningf(..., n = 1L)
}
\description{
Implements \code{warning} and \code{sprintf} such that the specified
environment (of a function) is properly shown in the message.
}
\details{
None.
}
