% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_dir}
\alias{create_dir}
\title{Create a directory.}
\usage{
create_dir(dir = paste0(getwd(), "/"), overwrite = FALSE, backup = TRUE,
  create = TRUE, ...)
}
\arguments{
\item{dir}{character. Folder name (with a trailing slash),}

\item{overwrite}{boolean.}

\item{backup}{boolean.}

\item{create}{boolean.}
}
\value{
character. Filepath.
}
\description{
Creates the directory if it does not exist.
}
\details{
Used by other functions in this package.
}
