% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create an example project/file.}
\usage{
create_example(style = "personal", template = "analysis",
  dir = "../project_example", ..., overwrite = TRUE)
}
\arguments{
\item{style}{Character. Default: "personal"}

\item{template}{Character. Default: "all"}

\item{dir}{Character. Name of directory to create. Default = "project_exmample"}

\item{...}{Dots. Parameters passed to ???}

\item{overwrite}{Boolean. Indicates whether to overwrite or not.}
}
\value{
Character. Directory filepath.
}
\description{
Creates an example project or file.
}
\details{
TODO.
}
