% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{do_call_with}
\alias{do_call_with}
\title{\code{do.call()} with package namespace}
\source{
\url{https://stackoverflow.com/questions/10022436/do-call-in-combination-with}.
}
\usage{
do_call_with(what, args, ...)
}
\arguments{
\item{what}{character. package name and function specified in explicit
namespace notation (i.e. \code{package::function}).}

\item{args}{list. Arguments passed to function parsed from \code{what}.}

\item{...}{dots. Arguments passed to \code{do.call()}.}
}
\description{
Allows a package to explicitly included in \code{do.call()}
}
\details{
None.
}
