% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_ext}
\alias{export_ext}
\alias{export_ext_csv}
\alias{export_ext_xlsx}
\alias{export_ext_rdata}
\alias{export_ext_RData}
\alias{export_ext_rda}
\alias{export_ext_rds}
\alias{export_ext_png}
\alias{export_gg}
\alias{export_viz}
\title{Save to a path}
\usage{
export_ext(x = NULL, basename = deparse(substitute(x)), dir = getwd(),
  ext = NULL, path = file.path(dir, paste0(basename, ".", ext)),
  overwrite = TRUE, backup = FALSE, path_backup = NULL, export = TRUE,
  return = TRUE, ...)

export_ext_csv(...)

export_ext_xlsx(...)

export_ext_rdata(...)

export_ext_RData(...)

export_ext_rda(...)

export_ext_rds(...)

export_ext_png(...)

export_gg(...)

export_viz(...)
}
\arguments{
\item{x}{data.frame (or matrix) for most formats.}

\item{basename}{character. Bare basename (i.e. without folderor extension),}

\item{dir}{character. Folder name (with a trailing slash).}

\item{ext}{character. Bare extension (i.e. without a dot). Must be one of valid formats.}

\item{path}{character. Concatenation of \code{basename}, \code{dir}, and \code{ext},}

\item{overwrite}{boolean.}

\item{backup}{boolean.}

\item{path_backup}{like \code{path},}

\item{export}{boolean. Indicates whether to actually carry out action. Intended to be used as a "catch all".}

\item{return}{boolean.  Relevant ONLY if \code{export == FALSE}.
Set to \code{TRUE} in order to preview what would be rendered.}

\item{...}{dots. Parameters to pass on to internally used export function.
Important for explicitly specifying non-default \code{ggplot2::ggsave()}.}
}
\value{
character. Filepath.
}
\description{
Saves an object to a path.
}
\details{
Object to save must be a data.frame (or matrix) for most formats.
}
\examples{
\dontrun{
if(interactive()){
 # These may be defined elsewhere in a project.
 dir_csv <- "csvs/"
 export_csv <- TRUE
 # ...
 export_ext(iris, ext = "csv", dir = dir_csv, export = export_csv)
 # Or...
 export_ext_csv(iris, dir = dir_csv, export = export_csv)
 }
}
}
