% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_path}
\alias{export_path}
\title{Export an object}
\usage{
export_path(x = NULL, path = NULL, overwrite = TRUE, backup = FALSE,
  path_backup = NULL, export = TRUE, return = TRUE, ...)
}
\arguments{
\item{x}{data.frame (or matrix) for most formats.}

\item{path}{character. Concatenation of \code{file}, \code{dir}, and \code{ext},}

\item{overwrite}{logical.}

\item{backup}{logical.}

\item{path_backup}{like \code{path},}

\item{export}{logical. Indicates whether to actually carry ret action. Intended to be used as a "catch all".}

\item{return}{logical. Relevant ONLY if \code{export == FALSE}.
Set to \code{TRUE} in order to preview what would be rendered.}

\item{...}{dots. Parameters to pass directly to the internally used export function.
Important for explicitly specifying non-default \code{ggplot2::ggsave()} parameters,
as well as specifying worksheets, etc. for an Excel-based export method.}
}
\value{
object.
}
\description{
Saves data given a full path.
}
\details{
Works similarly to \code{export_ext()} internally, but may be considered simpler.
}
