% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_path_safely}
\alias{get_path_safely}
\title{Construct a file path}
\usage{
get_path_safely(dir = NULL, file = NULL, ext = NULL, path = NULL)
}
\arguments{
\item{dir}{character. Folder name (with a trailing slash).}

\item{file}{character. Bare file (i.e. without folderor extension),}

\item{ext}{character. Bare extension (i.e. without a dot). Must be one of valid formats.}

\item{path}{character. Concatenation of \code{file}, \code{dir}, and \code{ext},}
}
\value{
character. Path.
}
\description{
Construct a file path given a directory, file, and extension.
}
\details{
This function is notable because it has a \code{path} argument
that can be used as a quick "out".
This helps with argument checking with package functions.
}
