% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ext}
\alias{import_ext}
\alias{import_ext_csv}
\alias{import_ext_xlsx}
\alias{import_ext_rdata}
\alias{import_ext_rda}
\alias{import_ext_rds}
\title{Import an object}
\usage{
import_ext(file, dir = getwd(), ext, path = file.path(dir, paste0(file, ".",
  ext)), ..., import = TRUE, return = TRUE)

import_ext_csv(...)

import_ext_xlsx(...)

import_ext_rdata(...)

import_ext_rda(...)

import_ext_rds(...)
}
\arguments{
\item{file}{character. Bare file (i.e. without folderor extension),}

\item{dir}{character. Folder name (with a trailing slash).}

\item{ext}{character. Bare extension (i.e. without a dot). Must be one of valid formats.}

\item{path}{character. Concatenation of \code{file}, \code{dir}, and \code{ext},}

\item{...}{dots. Arguments to pass dircetly to internally used import function.}

\item{import}{logical. Indicates whether to actually execute function.}

\item{return}{logical. Relevant ONLY if \code{export == FALSE}.
Set to \code{TRUE} in order to preview what would be rendered.}
}
\value{
object.
}
\description{
Reads in data given a file, directory, and extension, or a full path directly.
}
\details{
This function is intended to be a direct counterpart to the similarly
named export function. It should be used in a NSE manner. However, this may not be what the user wants.
(The NSE aspect is perhaps more appropriate for exporting, where choices about file name
may be determined from the object name.)
\code{import_path()} may be a suitable alternative.

Note that the user may not typically be concerned with the \code{import} and
\code{return} parameters. Nonetheless, these are provided in order to facilitate usage
with scripts run using "meta"-parameters that determine what data to import.

Internally, a distinct method (\code{session::restore.session})
is used for the RData-type file extension in order
to allow for importing of packages.

Many supplementary functions, using the format \code{import_ext_*()} are provided
for convenience. (e.g. \code{import_ext_csv()} instead of \code{import_ext(..., ext = 'csv').}
}
