% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ext}
\alias{import_ext}
\alias{import_ext_csv}
\alias{import_ext_xlsx}
\alias{import_excel}
\alias{import_ext_rdata}
\alias{import_ext_RData}
\alias{import_ext_rda}
\alias{import_ext_rds}
\title{Import an object.}
\usage{
import_ext(x = NULL, filename = deparse(substitute(x)),
  dir = paste0(getwd(), "/"), ext = NULL, filepath = paste0(dir, filename,
  ".", ext), import = TRUE, ...)

import_ext_csv(...)

import_ext_xlsx(...)

import_excel(...)

import_ext_rdata(...)

import_ext_RData(...)

import_ext_rda(...)

import_ext_rds(...)
}
\arguments{
\item{x}{data.frame (or matrix) for most formats.}

\item{filename}{character. Bare filename (i.e. without folderor extension),}

\item{dir}{character. Folder name (with a trailing slash).}

\item{ext}{character. Bare extension (i.e. without a dot). Must be one of valid formats.}

\item{filepath}{character. Concatenation of `filename`, `dir`, and \code{ext},}

\item{import}{boolean. Indicates whether to actually carry out function.}

\item{...}{dots. May (or may not) be passed on to specific exporting functions.}
}
\description{
Reads in an object from a filepath.
}
\details{
None.
}
