% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ext}
\alias{import_ext}
\alias{import_ext_csv}
\alias{import_ext_xlsx}
\alias{import_excel}
\alias{import_ext_rdata}
\alias{import_ext_RData}
\alias{import_ext_rda}
\alias{import_ext_rds}
\title{Import an object}
\usage{
import_ext(filename, dir = getwd(), ext, filepath = file.path(dir,
  paste0(filename, ".", ext)), import = TRUE, return = TRUE, ...)

import_ext_csv(...)

import_ext_xlsx(...)

import_excel(...)

import_ext_rdata(...)

import_ext_RData(...)

import_ext_rda(...)

import_ext_rds(...)
}
\arguments{
\item{filename}{character. Bare filename (i.e. without folderor extension),}

\item{dir}{character. Folder name (with a trailing slash).}

\item{ext}{character. Bare extension (i.e. without a dot). Must be one of valid formats.}

\item{filepath}{character. Concatenation of \code{filename}, \code{dir}, and \code{ext},}

\item{import}{boolean. Indicates whether to actually carry out function.}

\item{return}{boolean.  Relevant ONLY if \code{export == FALSE}.
Set to \code{TRUE} in order to preview what would be rendered.}

\item{...}{dots. Parameters to pass on to internally used import function.}
}
\description{
Reads in an object from a filepath.
}
\details{
None.
}
