% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_proj_io.R
\name{parse_proj_io}
\alias{parse_proj_io}
\title{Parse input and output in project scripts.}
\usage{
parse_proj_io(filepaths, dir, ..., rgx_file_io = ".R",
  rgx_input = ".*(import_ext|teutils::import_ext)",
  rgx_output = ".*(export_ext|teutils::export_ext)")
}
\arguments{
\item{filepaths}{character. Can be a vector. Missing by default. Takes precedence over \code{dir} argument.}

\item{dir}{character. Should not be a vector. Missing by default. Only used if \code{filepaths} is missing.}

\item{...}{dots. Arguments passed to \code{list.files()} for identifying input/output files.
(Technically, \code{pattern} could/should not be specified explicitly.)}

\item{rgx_file_io}{character. Alias to `pattern` parameter for `list.files()`. Used ONLY if `filepaths` is missing and \code{dir} is not.}

\item{rgx_input}{character. Regular expression to match for input functions}

\item{rgx_output}{Chracter. Regular expression to match for output functions.}
}
\value{
data.frame
}
\description{
Parses project diretory to identify input andoutput variables, filenames, etc.
}
\details{
Intended to be used with this package's `export*()` and \code{import*()} functions, but it can also work
with the functions from other packages (e.g. `readr`) if `rgx_input` and \code{rgx_output} are modified appropriately.
For example, `rgx_input == "read"` and `rgx_output == "write"` might be appropriate choices if using the \code{readr} package.)
Does not (currently) work as well if variables, filenames, etc. are not wrapped directly by
the functions described by `rgx_input` and \code{rgx_output}.
}
