% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_proj_io.R
\name{render_proj_io}
\alias{render_proj_io}
\title{Render project output.}
\usage{
render_proj_io(filepaths_input, dir_input, ..., dir_output = NULL,
  filenames_output, rgx_input = ".", rgx_input_include = rgx_input,
  rgx_input_exclude = "", rgx_output_trim = "-v[0-9]+.R",
  ext_output = ".html", render = TRUE, rgx_render = ".",
  overwrite = TRUE, backup = FALSE, backup_suffix = paste0("_",
  format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S")), keep_rmd = FALSE)
}
\arguments{
\item{filepaths_input}{character}

\item{dir_input}{character}

\item{...}{character}

\item{dir_output}{character. File path.}

\item{filenames_output}{character.}

\item{rgx_input}{character.}

\item{rgx_input_include}{character.}

\item{rgx_input_exclude}{character.}

\item{rgx_output_trim}{character.}

\item{ext_output}{character.}

\item{render}{character.}

\item{rgx_render}{character}

\item{overwrite}{boolean.}

\item{backup}{character.}

\item{backup_suffix}{character.}

\item{keep_rmd}{character.}
}
\value{
data.frame.
}
\description{
Parses project to identify inputs/out variables, filenames, etc.
}
\details{
None.
}
