% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ext.r
\name{set_pkg_print_opts}
\alias{set_pkg_print_opts}
\title{Control package messages.}
\usage{
set_pkg_print_opts(msg = getOption("teproj.print_msg"),
  wrn = getOption("teproj.print_wrn"), err = getOption("teproj.print_err"))
}
\arguments{
\item{msg}{Boolean. Default: FALSE (getOption("teproj.print_msg")).}

\item{wrn}{Boolean. Default: TRUE (getOption("teproj.print_wrn")).}

\item{err}{Boolean. Default: TRUE (getOption("teproj.print_err"))}
}
\description{
Sets print message parameters for `teproj` functions.
}
\details{
Intended to be used as a wrapper to `options(...`).
}
