% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_opts.R
\name{set_pkg_print_opts}
\alias{set_pkg_print_opts}
\title{Control package options}
\usage{
set_pkg_print_opts(msg = getOption("teproj.print.msg"),
  wrn = getOption("teproj.print.wrn"), err = getOption("teproj.print.err"))
}
\arguments{
\item{msg, wrn, err}{booleans. Indiciates whether to show messages, warnings, and errors for package functions.}
}
\description{
Sets options for relevant \code{teproj} functions.
}
\details{
Intended to be used as a wrapper to \code{options(...)}.
}
